1 file suffices
